show score increase in GameController

+++ src/main/java/koeko/controllers/Game/StudentCellView.kt
@@ -4,9 +4,9 @@ import koeko.students_management.Student
 
 class StudentCellView {
     var student: Student = Student()
-    var score = 0
+    var score = 0.0
 
-    constructor(student: Student, initialScore: Int = 0) {
+    constructor(student: Student, initialScore: Double = 0.0) {
         this.student = student
         this.score = initialScore
     }
