+++ src/main/java/koeko/controllers/Game/Team.kt
@@ -4,9 +4,9 @@ import koeko.students_management.Student
 
 class Team {
     var studentCellView: ArrayList<StudentCellView> = ArrayList()
-    var teamScore = 0
+    var teamScore = 0.0
 
-    fun increaseScore (student: StudentCellView, scoreIncrease: Int) {
+    fun increaseScore (student: StudentCellView, scoreIncrease: Double) {
         var index = studentCellView.indexOf(student)
         if (index >= 0) {
             teamScore += scoreIncrease
